refactor: use data output period types hook
