+++ src/components/PeriodDimension/PeriodDimension.js
@@ -1,8 +1,8 @@
 import { useConfig, useDataQuery } from '@dhis2/app-runtime'
 import PropTypes from 'prop-types'
-import React, { useEffect, useMemo } from 'react'
 import { DIMENSION_ID_PERIOD } from '../../modules/predefinedDimensions.js'
 import PeriodTransfer from './PeriodTransfer.js'
+import { useDataOutputPeriodTypes } from './useDataOutputPeriodTypes.js'
 
 const userSettingsQuery = {
     userSettings: {
@@ -11,26 +11,6 @@ const userSettingsQuery = {
             key: ['keyUiLocale'],
         },
     },
-}
-
-const v43Query = {
-    enabledPeriodTypes: {
-        resource: 'configuration/dataOutputPeriodTypes',
-    },
-    financialYearStart: {
-        resource: 'systemSettings/analyticsFinancialYearStart',
-    },
-    analysisRelativePeriod: {
-        resource: 'systemSettings/keyAnalysisRelativePeriod',
-    },
-}
-
-const FY_SETTING_TO_SERVER_PT = {
-    FINANCIAL_YEAR_APRIL: 'FinancialApril',
-    FINANCIAL_YEAR_JULY: 'FinancialJuly',
-    FINANCIAL_YEAR_SEPTEMBER: 'FinancialSep',
-    FINANCIAL_YEAR_OCTOBER: 'FinancialOct',
-    FINANCIAL_YEAR_NOVEMBER: 'FinancialNov',
 }
 
 const SELECTED_PERIODS_PROP_DEFAULT = []
@@ -44,80 +24,17 @@ const PeriodDimension = ({
     height,
 }) => {
     const config = useConfig()
-    const { systemInfo, serverVersion } = config
+    const { systemInfo } = config
     const userSettingsResult = useDataQuery(userSettingsQuery)
 
-    const supportsEnabledPeriodTypes = serverVersion.minor >= 43
-
-    const {
-        data: v43Data,
-        error: v43Error,
-        refetch: v43Refetch,
-    } = useDataQuery(v43Query, { lazy: true })
-
-    useEffect(() => {
-        if (supportsEnabledPeriodTypes) {
-            v43Refetch()
-        }
-    }, [supportsEnabledPeriodTypes, v43Refetch])
+    const { supportsEnabledPeriodTypes, enabledPeriodTypesData } =
+        useDataOutputPeriodTypes()
 
     const { calendar = 'gregory' } = systemInfo
     const { data: { userSettings: { keyUiLocale: locale } = {} } = {} } =
         userSettingsResult
 
     const periodsSettings = { calendar, locale }
-
-    const enabledPeriodTypesData = useMemo(() => {
-        if (!supportsEnabledPeriodTypes) {
-            return null
-        }
-
-        if (v43Error || !v43Data?.enabledPeriodTypes) {
-            return null
-        }
-
-        const enabledTypes = v43Data.enabledPeriodTypes
-
-
-        if (!enabledTypes || enabledTypes.length === 0) {
-            return {
-                enabledTypes: [],
-                financialYearStart: null,
-                analysisRelativePeriod: null,
-                noEnabledTypes: true,
-            }
-        }
-
-        let financialYearStart = null
-        let financialYearDisplayLabel = null
-        if (v43Data.financialYearStart?.analyticsFinancialYearStart) {
-            const fyStartValue =
-                v43Data.financialYearStart.analyticsFinancialYearStart
-
-            const mappedFyPt = FY_SETTING_TO_SERVER_PT[fyStartValue]
-            const matchingPt = enabledTypes.find(
-                (pt) => pt.name === mappedFyPt
-            )
-            matchingPt.displayLabel = 'Academic year'
-            if (mappedFyPt && matchingPt) {
-                financialYearStart = fyStartValue
-                if (matchingPt.displayLabel) {
-                    financialYearDisplayLabel = matchingPt.displayLabel
-                }
-            }
-        }
-
-        const analysisRelativePeriod =
-            v43Data.analysisRelativePeriod?.keyAnalysisRelativePeriod || null
-
-        return {
-            enabledTypes,
-            financialYearStart,
-            financialYearDisplayLabel,
-            analysisRelativePeriod,
-            noEnabledTypes: false,
-        }
-    }, [supportsEnabledPeriodTypes, v43Data, v43Error])
 
     const selectPeriods = (periods) => {
         onSelect({
