+++ src/index.js
@@ -5,6 +5,7 @@ export { default as DataDimension } from './components/DataDimension/DataDimensi
 
 export { default as PeriodDimension } from './components/PeriodDimension/PeriodDimension.js'
 export { default as FixedPeriodSelect } from './components/PeriodDimension/FixedPeriodSelect.js'
+export { useDataOutputPeriodTypes } from './components/PeriodDimension/useDataOutputPeriodTypes.js'
 
 export { default as OrgUnitDimension } from './components/OrgUnitDimension/OrgUnitDimension.js'
